Rename role id complementation mapper class
